requirements.txt
# Основные пакеты для работы с aiogram
aiogram
aiohttp[speedups]==3.9.5
aiodns
aiofiles==24.1.0
ujson
colorama
fluent.runtime
fluent.syntax
pydantic==2.9.2
pydantic_settings==2.1.0
structlog==24.4.0

# Дополнительные async-библиотеки для aiogram
aiosqlite  # Async work with SQLite
sqlalchemy[asyncio]==2.0.36

# Логирование, интернационализация и прочие вспомогательные пакеты
annotated-types==0.7.0
attrs==24.2.0
Babel==2.16.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
magic-filter==1.0.12
multidict==6.1.0
packaging==24.1
pluggy==1.5.0
propcache==0.2.0
pycares==4.4.0
pycparser==2.22
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
ruff==0.7.1
tomli==2.0.1
typing_extensions==4.12.2

# Пакеты для тестирования
pytest
pytest-asyncio
pytest-cov
coverage

# Дополнительные пакеты для работы с OpenAI
openai>=1.0.0
aiohttp>=3.9.5
